latest_ai_development/pyproject.toml
[project]
name = "latest_ai_development"
version = "0.1.0"
description = "latest-ai-development using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai-tools>=1.7.0",
    "crewai[google-genai,tools]==1.7.0",
    "litellm>=1.75.3",
    "fastapi",
    "apscheduler",
    # Observability - Langfuse integration
    "langfuse>=2.0.0",
    "openinference-instrumentation-crewai>=0.1.0",
]

[project.scripts]
latest_ai_development = "latest_ai_development.main:run"
run_crew = "latest_ai_development.main:run"
train = "latest_ai_development.main:train"
replay = "latest_ai_development.main:replay"
test = "latest_ai_development.main:test"
run_with_trigger = "latest_ai_development.main:run_with_trigger"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
